Made components more generic. Moved fetch out of the main page so that that page get return to a server component.
