+++ src/components/ui/data-table/data-table.tsx
@@ -1,0 +1,108 @@
+"use client"
+
+import { ReactNode } from "react"
+import {
+  ColumnDef,
+  flexRender,
+  getCoreRowModel,
+  useReactTable,
+  TableOptions,
+  Table as TanstackTable,
+} from "@tanstack/react-table"
+
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@/components/ui/table"
+
+interface DataTableProps<TData, TValue> {
+  columns: ColumnDef<TData, TValue>[]
+  data: TData[]
+  tableOptions?: Partial<TableOptions<TData>>
+  renderFilters?: (table: TanstackTable<TData>) => ReactNode
+  renderPagination?: (table: TanstackTable<TData>) => ReactNode
+}
+
+const DataTable = <TData, TValue>({
+  columns,
+  data,
+  tableOptions = {},
+  renderFilters,
+  renderPagination,
+}: DataTableProps<TData, TValue>) => {
+  const table = useReactTable({
+    data,
+    columns,
+    getCoreRowModel: getCoreRowModel(),
+    ...tableOptions,
+  })
+
+  return (
+    <div className="flex flex-col h-full">
+      {renderFilters && <div className="mt-2">{renderFilters(table)}</div>}
+      <div className="rounded-md border flex-grow">
+        <Table>
+          <TableHeader>
+            {table.getHeaderGroups().map((headerGroup) => (
+              <TableRow key={headerGroup.id}>
+                {headerGroup.headers.map((header) => (
+                  <TableHead
+                    key={header.id}
+                    style={{
+                      width: header.column.getSize(),
+                      maxWidth: header.column.getSize(),
+                    }}
+                  >
+                    {header.isPlaceholder
+                      ? null
+                      : flexRender(
+                          header.column.columnDef.header,
+                          header.getContext()
+                        )}
+                  </TableHead>
+                ))}
+              </TableRow>
+            ))}
+          </TableHeader>
+          <TableBody>
+            {table.getRowModel().rows?.length ? (
+              table.getRowModel().rows.map((row) => (
+                <TableRow
+                  key={row.id}
+                  data-state={row.getIsSelected() && "selected"}
+                >
+                  {row.getVisibleCells().map((cell) => (
+                    <TableCell key={cell.id} className="align-top">
+                      {flexRender(
+                        cell.column.columnDef.cell,
+                        cell.getContext()
+                      )}
+                    </TableCell>
+                  ))}
+                </TableRow>
+              ))
+            ) : (
+              <TableRow>
+                <TableCell
+                  colSpan={columns.length}
+                  className="h-24 text-center"
+                >
+                  No results.
+                </TableCell>
+              </TableRow>
+            )}
+          </TableBody>
+        </Table>
+      </div>
+      {renderPagination && (
+        <div className="mt-4">{renderPagination(table)}</div>
+      )}
+    </div>
+  )
+}
+
+export { DataTable }
